requirements.txt
opencv-python
numpy
imageio[ffmpeg]
cupy-cuda12x
